Review comment Max v1.0
