+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -81,7 +81,6 @@ iree_compiler_cc_library(
         "BufferizeDispatchTensorLoadStore.cpp",
         "CleanupBufferAllocViewPass.cpp",
         "CombineLayoutTransformation.cpp",
-        "CombineLayoutTransformationForMapGather.cpp",
         "ConcretizePadResultShape.cpp",
         "ConfigTrackingCanonicalizer.cpp",
         "ConvertAccGEMMToGEMMPass.cpp",
