+++ compiler/src/iree/compiler/Codegen/Common/test/BUILD.bazel
@@ -29,8 +29,8 @@ iree_lit_test_suite(
             "canonicalize_early_bufferization_ops.mlir",
             "canonicalize_interface_load_store.mlir",
             "check_for_config.mlir",
-            "combine_layout_transformation.mlir",
-            "combine_layout_transformation_for_map_gather.mlir",
+            "combine_result_layout_transformation.mlir",
+            "combine_source_layout_transformation.mlir",
             "convert_accgemm_to_gemm.mlir",
             "convert_bf16_arith_to_f32.mlir",
             "convert_bf16_to_uint16_buffers.mlir",
