+++ compiler/src/iree/compiler/Codegen/LLVMCPU/Passes.cpp
@@ -133,7 +133,7 @@ addTileAndDistributePasses(OpPassManager &funcPassManager,
     funcPassManager.addPass(
         createTileAndDistributeToWorkgroupsUsingForallOpPass());
     funcPassManager.addPass(createBufferizeDispatchTensorLoadStorePass());
-    funcPassManager.addPass(createCombineLayoutTransformationPass());
+    funcPassManager.addPass(createCombineResultLayoutTransformationPass());
   } else {
     funcPassManager.addPass(createTileAndDistributeToWorkgroupsPass());
     funcPassManager.addPass(createCSEPass());
