+++ compiler/src/iree/compiler/Codegen/LLVMGPU/Passes.cpp
@@ -538,11 +538,11 @@ void addGPUTileAndFusePassPipeline(OpPassManager &funcPassManager,
 
   // Step 5. Greedily fuse parallel loops and hoist from serial loops.
   funcPassManager.addPass(createGPUFuseAndHoistParallelLoopsPass());
-  CombineLayoutTransformationPassOptions combineLayoutOptions;
+  CombineResultLayoutTransformationPassOptions combineLayoutOptions;
   combineLayoutOptions.scope =
       IREE::Codegen::RelayoutCombinationScope::Workgroup;
   funcPassManager.addPass(
-      createCombineLayoutTransformationPass(combineLayoutOptions));
+      createCombineResultLayoutTransformationPass(combineLayoutOptions));
   funcPassManager.addPass(createGPUGreedilyDistributeToThreadsPass());
   funcPassManager.addPass(createTileLargeTensorsPass());
   funcPassManager.addPass(createCanonicalizerPass());
