+++ compiler/src/iree/compiler/Dialect/LinalgExt/IR/BUILD.bazel
@@ -104,6 +104,7 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:TensorDialect",
         "@llvm-project//mlir:TensorUtils",
         "@llvm-project//mlir:TilingInterface",
+        "@llvm-project//mlir:UBDialect",
         "@llvm-project//mlir:ValueBoundsOpInterface",
         "@llvm-project//mlir:ViewLikeInterface",
     ],
