+++ compiler/src/iree/compiler/Dialect/LinalgExt/IR/CMakeLists.txt
@@ -65,6 +65,7 @@ iree_cc_library(
     MLIRTensorDialect
     MLIRTensorUtils
     MLIRTilingInterface
+    MLIRUBDialect
     MLIRValueBoundsOpInterface
     MLIRViewLikeInterface
     iree::compiler::Dialect::LinalgExt::Utils
